feat: 이미지 다운로드 중복 방지 기능 구현
