+++ NaeilBurgerCamp/NaeilBurgerCamp/Utility/Extensions/UIImageView+Extensions.swift
@@ -3,7 +3,7 @@ import UIKit
 extension UIImageView {
     func setImage(with urlString: String) {
         Task {
-            let image = await ImageCacheManger.shared.loadImage(for: urlString)
+            let image = await ImageLoader.shared.loadImage(for: urlString)
             await MainActor.run {
                 self.image = image
             }
